pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bidmate-rag"
version = "0.1.0"
description = "RAG system scaffold for public and enterprise RFP analysis."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.6",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "hwp-hwpx-parser>=1.0.0",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.3.3",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
bidmate-eval = "bidmate_rag.cli.eval:main"
bidmate-report = "bidmate_rag.cli.report:main"
bidmate-compare = "bidmate_rag.cli.compare:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.5",
]
notebook = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
]
ui = [
    "streamlit",
]
ml = [
    "datasets",
    "peft",
    "sentence-transformers",
    "transformers",
    "trl",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bidmate_rag"]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["experiments/notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E501",  # line-too-long: 한국어 주석/메시지가 100자 자주 초과
    "E402",  # module-import-not-at-top: load_dotenv()를 import 전에 호출하는 패턴
]
